RAC-342 refactor : 예외 메시지 및 코드 패키지 변경
