+++ src/test/java/com/postgraduate/domain/mentoring/application/usecase/MentoringSeniorInfoUseTypeTest.java
@@ -39,7 +39,7 @@ import static org.mockito.BDDMockito.given;
 import static org.mockito.BDDMockito.mock;
 
 @ExtendWith(MockitoExtension.class)
-class MentoringSeniorInfoUseCaseTest {
+class MentoringSeniorInfoUseTypeTest {
     @Mock
     private MentoringGetService mentoringGetService;
     @Mock
