+++ src/test/java/com/postgraduate/domain/salary/application/usecase/SalaryRenewalUseTypeTest.java
@@ -15,7 +15,7 @@ import java.time.LocalDate;
 import static org.mockito.Mockito.*;
 
 @ExtendWith(MockitoExtension.class)
-class SalaryRenewalUseCaseTest {
+class SalaryRenewalUseTypeTest {
     @Mock
     private SalarySaveService salarySaveService;
     @InjectMocks
