+++ src/test/java/com/postgraduate/domain/senior/application/usecase/SeniorInfoUseTypeTest.java
@@ -35,7 +35,7 @@ import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 
 @ExtendWith(MockitoExtension.class)
-class SeniorInfoUseCaseTest {
+class SeniorInfoUseTypeTest {
     @Mock
     private SeniorGetService seniorGetService;
     @Mock
