+++ src/test/java/com/postgraduate/domain/senior/application/usecase/SeniorMyPageUseTypeTest.java
@@ -42,7 +42,7 @@ import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.mock;
 
 @ExtendWith(MockitoExtension.class)
-class SeniorMyPageUseCaseTest {
+class SeniorMyPageUseTypeTest {
     @Mock
     private SeniorGetService seniorGetService;
     @Mock
